Conflicts:
	src/main/java/it/tredi/ecm/dao/entity/Evento.java
	src/main/resources/i18n/messages.properties
	src/main/resources/static/css/main.css
	src/main/resources/templates/evento/eventoRESEdit.html

+++ src/main/java/it/tredi/ecm/dao/entity/DettaglioAttivitaFAD.java
@@ -28,6 +28,6 @@ public class DettaglioAttivitaFAD implements Serializable{
 	private ObiettiviFormativiFADEnum obiettivoFormativo;
 	@Enumerated(EnumType.STRING)
 	private MetodologiaDidatticaFADEnum metodologiaDidattica;
-	private Long oreAttivita;
+	private float oreAttivita;
 }
